+++ src/LASRstages/writelas.cpp
@@ -57,11 +57,19 @@ bool LASRlaswriter::process(LASpoint*& p)
 {
   if (p->get_withheld_flag() != 0) return true;
 
+  // No writer initialized? Create a writer.
   if (!laswriter)
   {
     LASwriteOpener laswriteopener;
     laswriteopener.set_file_name(ofile.c_str());
     laswriter = laswriteopener.open(lasheader);
+
+    offsets[0] = p->quantizer->x_offset;
+    offsets[1] = p->quantizer->y_offset;
+    offsets[2] = p->quantizer->z_offset;
+    scales[0] = p->quantizer->x_scale_factor;
+    scales[1] = p->quantizer->y_scale_factor;
+    scales[2] = p->quantizer->z_scale_factor;
 
     if (!laswriter)
     {
@@ -72,10 +80,36 @@ bool LASRlaswriter::process(LASpoint*& p)
     written.push_back(ofile);
   }
 
+  //  If the point in not in the buffer we can write it
   if (keep_buffer || !p->inside_buffer(xmin, ymin, xmax, ymax, circular))
   {
     if (!lasfilter.filter(p))
     {
+      //print("Before %d %d %d \n", p->get_X(), p->get_Y(), p->get_Z());
+
+      // If we write in a merged file the points may come from different file formats
+      if (merged)
+      {
+        if (p->quantizer->x_offset != offsets[0] || p->quantizer->x_scale_factor != scales[0])
+        {
+          double coordinate = (p->get_x() - offsets[0])/scales[0];
+          p->set_X(I32_QUANTIZE(coordinate));
+        }
+
+        if (p->quantizer->y_offset != offsets[1] || p->quantizer->y_scale_factor != scales[1])
+        {
+          double coordinate = (p->get_y() - offsets[1])/scales[1];
+          p->set_Y(I32_QUANTIZE(coordinate));
+        }
+
+        if (p->quantizer->z_offset != offsets[2] || p->quantizer->z_scale_factor != scales[2])
+        {
+          double coordinate = (p->get_z() - offsets[2])/scales[2];
+          p->set_Z(I32_QUANTIZE(coordinate));
+        }
+      }
+
+      //print("After %d %d %d \n\n", p->get_X(), p->get_Y(), p->get_Z());
       laswriter->write_point(p);
       laswriter->update_inventory(p);
     }